requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==23.0
pyrsistent==0.19.3
pytz==2022.7
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
